DisBot/requirements.txt
discord.py==2.6.4
Flask==3.1.1
python-dotenv==1.1.1
